Nova Landing page

+++ oportunidadesEmprego/src/componentes/PagRecrutador/Dashboard/Dashboard.jsx
@@ -18,20 +18,15 @@ export default function Dashboard() {
         fetch(`http://localhost:5000/minhasVagas/${currentUser.usuário.recruterEmail}`)
             .then((res) => res.json())
             .then(async (data) => {
-               
-               
-                // Loop through each job to fetch the number of applications
                 const jobsWithApplications = await Promise.all(data.map(async (job) => {
-                    
                     const response = await fetch(`http://localhost:5000/candidaturas/${job._id}`);
                     const applications = await response.json();
-                    
                     return { ...job, numApplications: applications.numCandidaturas };
                 }));
                 setJobs(jobsWithApplications);
                 setIsLoading(false);
             });
-    }, []);
+    }, [currentUser.usuário.recruterEmail]);
 
     const handleSearch = () => {
         const filter = jobs.filter((job) => job.jobTitle.toLowerCase().indexOf(searchText.toLowerCase()) !== -1);
@@ -40,18 +35,20 @@ export default function Dashboard() {
     };
 
     const handleDelete = (id) => {
-        fetch(`http://localhost:5000/vaga/${id}`, {
-            method: "DELETE"
-        })
-            .then(res => res.json())
-            .then(data => {
-                window.location.reload();
-            });
+        const confirmDelete = window.confirm("Tem certeza que deseja apagar esta vaga?");
+        if (confirmDelete) {
+            fetch(`http://localhost:5000/vaga/${id}`, {
+                method: "DELETE"
+            })
+                .then(res => res.json())
+                .then(data => {
+                    window.location.reload();
+                });
+        }
     };
 
     return (
-        <div className='Dashboard' >
-
+        <div className='Dashboard'>
             <section className='RecrutadorCandidatosDiv'>
                 <div>
                     <div className='SearchCandidatos'>
@@ -66,16 +63,15 @@ export default function Dashboard() {
                             {jobs.map((job, index) => (
                                 <div key={index}>
                                     <div>
-                                        <img src={job.companyLogo} />
+                                        <img src={job.companyLogo} alt="Logo da empresa" />
                                         <h2>{job.jobTitle}</h2>
                                     </div>
                                     <p className='EsspecialJob'>{job.location}</p>
-                                    {console.log(job.numApplications)}
                                     <p className='EsspecialJob'>Candidaturas: <span style={{ backgroundColor: '#652ebe', color: 'white', borderRadius: "20px", padding: "5px" }}>{job.numApplications}</span></p>
                                     <div>
                                         <button onClick={() => handleDelete(job._id)}>Apagar</button>
                                         <Link to={`/actualizar-vaga/${job._id}`}>Editar</Link>
-                                        <Link to={`/candidaturas-vaga/${job._id}`}> Candidaturas</Link>
+                                        <Link to={`/candidaturas-vaga/${job._id}`}>Candidaturas</Link>
                                     </div>
                                 </div>
                             ))}
@@ -86,3 +82,4 @@ export default function Dashboard() {
         </div>
     );
 }
+
